chore: Reorg types
