+++ app/client/packages/design-system/ads/src/Templates/Sidebar/SidebarButton/index.ts
@@ -1,1 +1,1 @@
-export { default } from "./SidebarButton";
+export { SidebarButton } from "./SidebarButton";
