+++ app/client/packages/design-system/ads/src/Templates/Sidebar/index.tsx
@@ -1,3 +1,3 @@
 export { IDESidebar } from "./Sidebar";
-export type { IDESidebarButton } from "./Sidebar";
 export { Condition } from "./enums";
+export { IDESidebarButton } from "./types";
